Data.Serialization

Description

headers

Contains an easily extensible Serializable class that allows converting an object from and to binary data. Versioning and dynamic typing are also handled, as well as writing to and reading from files and such.

Generic specifies generic versions of the serializer and deserializer that can be used for arbitrary data structures. It can be used in conjunction with custom Serializable instances if you want to manually introduce optimizations to the binary representation of (a partial) structure.

Types

data Serialized

Represents the binary representation of a serialized object, along with version and type information.

type LazyByteString = ByteString

Type definition that is useful to distinguish lazy from regular ByteStrings.

data SerializationException

Is thrown by the deserialization functions in case something goes wrong.

Constructors

IncompatibleLibraries
If an object has been serialized with another version of this library.

IncompatibleSerializers
In case a different version of the serializer for a particular object was used.

InvalidSerializedData
When the data read from a file or such does not conform to the protocol.

DeserializationError String

Serializable class

class Typeable a => Serializable a

Represents objects for which a specific serialization method has been defined. Instances exist for many standard types. If you want to serialize something of a different type, either use the generic serializers from Data.Serialization.Generic or make a custom instance of Serializable by importing Data.Serialization.Extent.

Serialization

serialize :: Serializable a => a -> IO Serialized

Encode a Serializable object into a binary representation that also contains type and version info. The resulting object can then be written to a file or stream using the storage functions.

deserialize :: Serializable a => Serialized -> IO (Maybe a)

Decodes a Serialized object back into a Haskell structure. In order to assert the type is correct, Nothing is returned when the inferred type is not correct. This works similarly as cast.

When a version incompatibility or other error occurs, a SerializationException is thrown.

Storage and retrieval of serialized objects

store :: FilePath -> Serialized -> IO ()

Writes a serialized object to a file.

load :: FilePath -> IO Serialized

Reads a serialized from a file.

stores :: FilePath -> [Serialized] -> IO ()

Writes multiple serialized objects to a file.

loads :: FilePath -> IO [Serialized]

Reads multiple serialized objects from a file.

hStore :: Handle -> Serialized -> IO ()

Writes a serialized object to a handle.

hStores :: Handle -> [Serialized] -> IO ()

Writes multiple serialized objects to a handle.

hLoad :: Handle -> IO Serialized

Loads a single serialized object from a handle. Note: this handle may not contain any data besides this objects.

hLoads :: Handle -> IO [Serialized]

Loads multiple serialized objects from a handle. Note: this handle may not contain any data besides these objects.

storeInByteString :: Serialized -> IO LazyByteString

Store the binary representation of a serialized object in a LazyByteString.

loadFromByteString :: LazyByteString -> IO (Serialized, LazyByteString)

Reads a serialized object from the beginning of a lazy bytestring. Returns it along with the rest of the string. Throws a SerializationException when an error occurs.