Data.Serialization.Extent

Description

headers

This module contains the definitions of the class methods of Serializable and therefore it can be used when you want to make a custom serializer for objects of a certain type.

Documentation

module Data.Serialization

type Byte = Word8

Byte is simply a synonym for a Word8.

data VersionID

A datatype containing a version of a serialization method.

Constructors

VersionID Int

class Typeable a => Serializable a where

Represents objects for which a specific serialization method has been defined. Instances exist for many standard types. If you want to serialize something of a different type, either use the generic serializers from Data.Serialization.Generic or make a custom instance of Serializable by importing Data.Serialization.Extent.

Methods

toBytes :: a -> [Byte]

Define an object's binary representation by converting it to a list of bytes. When implementing this, you do not have to take version and type information in account as that is taken care of by the library.

The only rule that should apply is that fromBytes (toBytes x) == x.

fromBytes :: [Byte] -> a

Decodes a binary representation to an object. You may assume that the argument is a valid result of toBytes.

serialVersionID :: a -> VersionID

A number indicating the version of the serialization method. This is included in the binary representation so an error can be generated whenever the serialization methods of two versions of a program are incompatible with each other.

This identifier should be incremented or otherwise changed whenever the definitions of toBytes and fromBytes have been updated.

The same ID should always be returned for any object of the same type, it should not depend on the contents of the argument (otherwise, undefined behavior may occur).

dependencies :: a -> [VersionID]

If the toBytes and fromBytes definitions for other Serializable types are used in order to encode an object of the current type. This should return a list of serialVersionID's of the serializers that are depended on.

Defining this function is optional and by default, it returns an empty list.

Like serialVersionID, the argument only indicates the type and it should furthermore be ignored.

constBytesSize :: a -> Maybe Int

Whenever the list returned by toBytes always has the same length, regardless of the input, this can be defined to indicate this constant size. When this is present, no lengths have to be encoded for this type, which can safe quite a few bytes, especially if many small objects are serialized.

Like serialVersionID, the argument only indicates the type and it should furthermore be ignored.

listToBytes :: [a] -> [Byte]

Can be implemented, in conjunction with its inverse listFromBytes to define a specialized serializer for lists of this object. For example, this is used to encode lists of Bool's as bitlists, making them eight times as small as with the default definition.

Is optional. If implemented, listFromBytes should be made compatible.

listFromBytes :: [Byte] -> [a]

See listToBytes.

sid :: Serializable a => a -> VersionID

Shorthand for serialVersionID. Is practical for when defining a lot of dependencies.