Data.Serialization.Generic

Description

headers

Defines generic serialization functions that require the type of objects that can serialized to be an instance of Data, rather than Serializable. The binary representation of objects serialized this way will depend on the layout of their data structure.

If objects of certain types need to be serialized in a certain way because doing so based on their layout is either not possible or less efficient, it is possible to manually make them an instance of Serializable and notifying the generic serializer of this by adding that type as a so-called specialization.

Whenever a data structure contains an object of a type for which a specialization exist, the serializer will use that for that object's binary representation; all objects inside the structure without such a specialization will be serialized using the generic method.

Documentation

Data.Serialization is exposed, with the exception of the functions serialize and deserialize. These have been given a different, generic, definition in this module.

module Data.Serialization

Generic serialization

serializeWith :: Data a => SerializationSettings -> a -> IO Serialized

Apply the generic serializer to some value, using certain settings. Functions from Serialization (such as store and load) can be used to handle the resulting Serialized object.

This function exists in the IO-monad because that is required for detection of sharing (multiple pointers referring to the same object).

serialize :: Data a => a -> IO Serialized

Serialize with default settings. See serializeWith and defaultSettings.

deserializeWith :: Data a => SerializationSettings -> Serialized -> IO (Maybe a)

Decodes a Serialized object back into a Haskell structure, using certain settings. In order to assert the type is correct, Nothing is returned when the inferred type is not correct.

The same SerializationSettings should be used as for the serialization of the object in question.

When a version incompatibility or other error occurs, a SerializationException is thrown.

deserialize :: Data a => Serialized -> IO (Maybe a)

Deserialize with default settings. See deserializeWith and defaultSettings.

Settings

data SerializationSettings

Holds the configurable settings for the serialization functions. This contains all the types for which specialized serializers exist.

Note that when an object x has been serialized with a :: SerializationSettings, trying to deserialize it with b :: SerializationSettings where a != b the possible incompatibility will be noticed and an exception will be thrown.

defaultSettings :: SerializationSettings

The default serialization settings. Contains specializations for basic types such as Int and Bool and most other types for which a Serializable-instance is defined in Extent.

Adding specializations

addSerializableSpecialization :: (Serializable a, Data a) => a -> SerializationSettings -> SerializationSettings

Add a serialization specialization of some type for which a Serializable instance exists and store it in the SerializationSettings. Only the type of the first argument is relevant, and it might just as well be undefined.

addSerSpec :: (Serializable a, Data a) => a -> SerializationSettings -> SerializationSettings

Shorthand for addSerializableSpecialization.