Data.Serialization.Internal.IntegralBytes

Description

This module offers basic functionality for converting integers to lists of bytes and vice-versa.

Documentation

bytes :: (Bits a, Integral a, Bounded a) => a -> [Word8]

Converts an integral to a list of bytes using a big endian byte order.

unbytes :: (Bits a, Integral a, Bounded a) => [Word8] -> a

Converts a list of bytes in big endian byte order to an integral that should be large enough to contain them.

varbytes :: (Bits a, Integral a) => a -> [Word8]

Uses a variable length encoding instead of a constant-sized one. Usefull for encoding lengths | that generally don't take up four or eight whole bytes and of which the size is | platform-specific. | | Currently only works on positive integers.

varunbytes :: (Bits a, Integral a) => [Word8] -> (a, [Word8])