Data.Serialization.Internal.Settings

Documentation

type ToByter m state = forall a. Data a => a -> state -> m ([Byte], state)

type FromByter m state = forall a. Data a => [Byte] -> state -> m (a, state)

class Data a => Serializable1 a where

Methods

toBytes1 :: Monad m => ToByter m s -> a -> s -> m ([Byte], s)

fromBytes1 :: Monad m => FromByter m s -> [Byte] -> s -> m (a, s)

serialVersionID1 :: a -> VersionID

data Serializer a

Constructors

Serializer (a -> [Byte]) ([Byte] -> a)

Serializer1 (forall m s. Monad m => ToByter m s -> a -> s -> m ([Byte], s)) (forall m s. Monad m => FromByter m s -> [Byte] -> s -> m (a, s))

NoSerializer

data SWrapper

Constructors

SWrapper (Generic Serializer) VersionID

data TypeKey

Constructors

TypeKey

Fields

unKey :: Either TypeRep TyCon

data SerializationSettings

Holds the configurable settings for the serialization functions. This contains all the types for which specialized serializers exist.

Note that when an object x has been serialized with a :: SerializationSettings, trying to deserialize it with b :: SerializationSettings where a != b the possible incompatibility will be noticed and an exception will be thrown.

Constructors

SerializationSettings

Fields

specializedInstances :: Map TypeKey SWrapper

settingsVID :: VersionID

typeKey :: TypeRep -> TypeKey

typeKey1 :: TypeRep -> TypeKey

assertType :: a -> a -> a

sWrapper :: (Serializable a, Data a) => a -> [(TypeKey, SWrapper)]

sWrapper1 :: Serializable1 (f a) => f a -> [(TypeKey, SWrapper)]

specializedSerializer :: Data a => SerializationSettings -> a -> Serializer a

standardSpecializations :: [(TypeKey, SWrapper)]

emptySettings :: SerializationSettings

defaultSettings :: SerializationSettings

The default serialization settings. Contains specializations for basic types such as Int and Bool and most other types for which a Serializable-instance is defined in Extent.

addSerializableSpecialization :: (Serializable a, Data a) => a -> SerializationSettings -> SerializationSettings

Add a serialization specialization of some type for which a Serializable instance exists and store it in the SerializationSettings. Only the type of the first argument is relevant, and it might just as well be undefined.

addSerSpec :: (Serializable a, Data a) => a -> SerializationSettings -> SerializationSettings

Shorthand for addSerializableSpecialization.